requirements.txt
asgiref==3.2.10
Django==3.1
astroid==2.0.4
certifi==2018.8.24
chardet==3.0.4
colorama==0.4.4
defusedxml==0.5.0
django-allauth==0.44.0
django-cors-headers==3.6.0
django-rest-auth==0.9.3
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.5.0
gunicorn==19.7.1
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauthlib==2.1.0
PyJWT==1.7.1
python3-openid==3.1.0
requests==2.21.0
requests-oauthlib==1.0.0
wrapt==1.10.11
beautifulsoup4==4.9.3
bs4==0.0.1
boto3==1.14.18
botocore==1.17.18
dj-database-url==0.5.0
django-braces==1.14.0
django-heroku==0.3.1
django-classy-tags==1.0.0
django-js-asset==1.2.2
django-s3direct==1.1.8
django-storages==1.9.1
django-tinymce==3.2.0
docutils==0.15.2
jmespath==0.10.0
Pillow==7.1.2
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
six==1.15.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
urllib3==1.24.3
whitenoise==5.1.0